fix -Wsign-compare

+++ hash_table8.hpp
@@ -1173,7 +1173,7 @@ public:
         if (required_buckets < _num_filled)
             return;
 
-        assert(required_buckets < max_size());
+        assert(required_buckets < (uint64_t)max_size());
         auto num_buckets = _num_filled > (1u << 16) ? (1u << 16) : 4u;
         while (num_buckets < required_buckets) { num_buckets *= 2; }
 #if EMH_SAVE_MEM
